Support streaming HTTP requests.
Currently, our round-tripper assumed that requests and response's bodies
were fully read within the function. It closed the stream (connection) as
soon as it returned.

This caused streaming requests to fail, as it managed to close the
connection before the response body was fully read.

This commit closes the connection only after the response's body is read
by the user.

It also writes the request in a separate go-routine, so that it does not
have to be fully written to start reading the response.